requirements.txt
alembic==1.8.1
blinker==1.5
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.40
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
